% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsprite2-package.R
\docType{package}
\name{rsprite2-package}
\alias{rsprite2}
\alias{rsprite2-package}
\title{rsprite2: Identify Distributions that Match Reported Sample Parameters (SPRITE)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The SPRITE algorithm creates possible distributions of discrete responses based on reported sample parameters, such as mean, standard deviation and range (Heathers et al., 2018, \doi{10.7287/peerj.preprints.26968v1}). This package implements it, drawing heavily on the code for Nick Brown's 'rSPRITE' Shiny app \url{https://shiny.ieis.tue.nl/sprite/}. In addition, it supports the modeling of distributions based on multi-item (Likert-type) scales and the use of restrictions on the frequency of particular responses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lukaswallrich.github.io/rsprite2/}
  \item Report bugs at \url{https://github.com/LukasWallrich/rsprite2/issues}
}

}
\author{
\strong{Maintainer}: Lukas Wallrich \email{lukas.wallrich@gmail.com} (\href{https://orcid.org/0000-0003-2121-5177}{ORCID})

Authors:
\itemize{
  \item Lukas Wallrich \email{lukas.wallrich@gmail.com} (\href{https://orcid.org/0000-0003-2121-5177}{ORCID})
}

Other contributors:
\itemize{
  \item Aurélien Allard \email{aurelien.ab.allard@gmail.com} [contributor]
  \item Lukas Jung \email{jung-lukas@gmx.net} [contributor]
}

}
\keyword{internal}
