% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{GRIM_test}
\alias{GRIM_test}
\title{GRIM test for mean}
\usage{
GRIM_test(mean, n_obs, m_prec = NULL, n_items = 1, return_values = FALSE)
}
\value{
Either TRUE/FALSE, or all possible means (if TRUE)/closest consistent mean (if FALSE)
}
\description{
This function tests whether a given mean (with a specific precision) can
result from a sample of a given size based on integer responses to one or more
items. The test is based on Brown & Heathers (2017).
If \code{return_values = TRUE} and if there is more than one precise mean compatible
with the given parameters, all possible means are returned. In that case, if the
given mean is not consistent, the closest consistent mean is returned with a
warning.
}
\details{
@inheritParams set_parameters mean n_obs m_prec, n_items
@param return_values Should all means consistent with the given parameters be returned?
}
\references{
\insertRef{brown2017grim}{rsprite2}
}
