% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{GRIM_test}
\alias{GRIM_test}
\title{GRIM test for mean}
\usage{
GRIM_test(
  mean,
  n_obs,
  m_prec = NULL,
  n_items = 1,
  return_values = FALSE,
  return_list = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{mean}{The mean of the distribution}

\item{n_obs}{The number of observations (sample size)}

\item{m_prec}{The precision of the mean, as number of digits after the decimal point.
If not provided, taken based on the significant digits of \code{mean} - so only needed if reported mean ends in 0}

\item{n_items}{Number of items in scale, if distribution represents scale averages.
Defaults to 1, which represents any single-item measure.}

\item{return_values}{A logical value. \emph{ignored if \code{return_list = TRUE}}.
If \code{FALSE} (the default), the function
returns a simple \code{TRUE} or \code{FALSE}. If \code{TRUE}, it returns a numeric
vector of all possible means that are consistent with the parameters.
If the test fails when \code{return_values = TRUE}, a warning is issued and
the closest consistent mean is returned. If there are two equally close
means, both are returned.}

\item{return_list}{A logical value. If \code{FALSE} (the default), the function's
return type is determined by \code{return_values}. If \code{TRUE}, the function
instead returns a list containing two elements: a logical \code{passed} flag
and a numeric vector \code{values} containing the relevant means.}

\item{quiet}{Suppress warnings.}
}
\value{
The return type depends on the arguments. By default, a logical scalar
(\code{TRUE} or \code{FALSE}). If \code{return_values = TRUE}, a numeric vector is returned.
If \code{return_list = TRUE}, a list is returned.
}
\description{
This function tests whether a given mean (with a specific precision) can
result from a sample of a given size based on integer responses to one or more
items. The test is based on Brown & Heathers (2017).
}
\examples{
# A sample of 28 integers cannot result in a mean of 5.19.
GRIM_test(mean = 5.19, n_obs = 28)

# To find the closest possible mean, set return_values to TRUE
GRIM_test(mean = 5.19, n_obs = 28, return_values = TRUE)

# To get a detailed list containing both the logical result and the values
GRIM_test(mean = 5.19, n_obs = 28, return_list = TRUE)

}
\references{
\insertRef{brown2017grim}{rsprite2}
}
