% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Plot distributions}
\usage{
plot_distributions(
  distributions,
  plot_type = c("auto", "histogram", "density"),
  max_plots = 100
)
}
\arguments{
\item{distributions}{Tibble with a column \code{distribution} and an identifier (\code{id}), typically as returned from \code{\link{find_possible_distributions}}}

\item{plot_type}{How many distributions should \emph{at most} be plotted? If more are passed, this number is randomly selected.}
}
\value{
A ggplot2 object that can be styled with functions such as \code{\link[ggplot2]{labs} or \code{\link[ggplot2]{theme_linedraw}}
}
\description{
This plots distributions identified by \code{\link{find_possible_distributions}} using ggplot2.
}
