% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsprite2-package.R
\docType{package}
\name{rsprite2-package}
\alias{rsprite2}
\alias{rsprite2-package}
\title{rsprite2: Identify distributions that match reported sample parameters (SPRITE)}
\description{
This package creates possible distributions based on reported sample
    parameters. For that, it uses the SPRITE algorithm proposed by Heathers 
    et al. (2018). Much of the code for is based on Nick Brown's rSPRITE shiny app,
    with some extensions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukasWallrich/rsprite2}
  \item Report bugs at \url{https://github.com/LukasWallrich/rsprite2/issues}
}

}
\author{
\strong{Maintainer}: Lukas Wallrich \email{lukas.wallrich@gmail.com} (\href{https://orcid.org/0000-0003-2121-5177}{ORCID})

}
\keyword{internal}
