% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsprite2-package.R
\docType{package}
\name{rsprite2-package}
\alias{rsprite2}
\alias{rsprite2-package}
\title{rsprite2: Identify Distributions that Match Reported Sample Parameters (SPRITE)}
\description{
The SPRITE algorithm creates possible distributions based on reported sample
    parameters (Heathers et al., 2018). This package implements it, drawing heavily on the
    code for for Nick Brown's rSPRITE shiny app. In addition, it supports the modeling
    of distributions based on multi-item (Likert-type) scales and the use of restrictions
    on the frequency of particular responses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukasWallrich/rsprite2}
  \item Report bugs at \url{https://github.com/LukasWallrich/rsprite2/issues}
}

}
\author{
\strong{Maintainer}: Lukas Wallrich \email{lukas.wallrich@gmail.com} (\href{https://orcid.org/0000-0003-2121-5177}{ORCID})

}
\keyword{internal}
